Rename 'features' to 'components'.
